feat: refactor Blockie components and add Button component with icons and animations
