--- packages/ui/src/Blockie/index.ts
@@ -1,2 +1,0 @@
-export { default as BlockieFace } from "./face";
-export { default as BlockieBottom } from "./bottom";
